+++ API/vitalapi/DTO'S/PlanoDto.cs
@@ -5,7 +5,7 @@
         public int Id { get; set; }
         public string Nome { get; set; }
         public string Descricao { get; set; }
-        public double Preco { get; set; }
+        public decimal Preco { get; set; }
         public string Tipo { get; set; }
         public string Funcionalidades { get; set; }
     }
@@ -13,13 +13,13 @@
     {
         public string Nome { get; set; }
         public string Descricao { get; set; }
-        public double Preco { get; set; }
+        public decimal Preco { get; set; }
         public string Tipo { get; set; }
         public string Funcionalidades { get; set; }
     }
     public class UpdatePlanoDto
     {
         public string Descricao { get; set; }
-        public double Preco { get; set; }
+        public decimal Preco { get; set; }
     }
 }
